Remove false positives: OU

+++ Cerebrum/OU.py
@@ -151,6 +151,26 @@ class OU(EntityContactInfo, EntityExternalId, EntityAddress,
             self._db.log_change(self.entity_id, self.clconst.ou_del, None)
         self.__super.delete()
 
+    def exists(self, table=None, binds=None):
+        """Check for existance"""
+        if not table or not binds:
+            raise ValueError('missing args')
+        exists_stmt = """
+        SELECT EXISTS (
+          SELECT 1
+          FROM [:table schema=cerebrum name={table}]
+          WHERE {where}
+        )
+        """.format(where=' AND '.join('{0}=:{0}'.format(x) for x in binds),
+                   table=table)
+        try:
+            self.query_1(exists_stmt, binds)
+            return True
+        except Errors.NotFoundError:
+            return False
+        except Errors.TooManyRowsError:
+            return True
+
     def find(self, ou_id):
         """Associate the object with the OU whose identifier is OU_ID.
 
@@ -210,12 +230,18 @@ class OU(EntityContactInfo, EntityExternalId, EntityAddress,
         return "/".join(components)
 
     def unset_parent(self, perspective):
-        self.execute("""
+        binds = {'ou_id': self.entity_id,
+                 'perspective': int(perspective)}
+        if not self.exists(table='ou_structure', binds=binds):
+            # false positive
+            return
+        delete_stmt = """
         DELETE FROM [:table schema=cerebrum name=ou_structure]
-        WHERE ou_id=:e_id AND perspective=:perspective""",
-                     {'e_id': self.entity_id,
-                      'perspective': int(perspective)})
-        self._db.log_change(self.entity_id, self.clconst.ou_unset_parent, None,
+        WHERE ou_id=:ou_id AND perspective=:perspective"""
+        self.execute(delete_stmt, binds)
+        self._db.log_change(self.entity_id,
+                            self.clconst.ou_unset_parent,
+                            None,
                             change_params={'perspective': int(perspective)})
 
     def set_parent(self, perspective, parent_id):
